+++ .idea/misc.xml
@@ -37,6 +37,7 @@
         <entry key="..\:/Users/fatim/AndroidStudioProjects/listary/app/src/main/res/layout/activity_new_list.xml" value="0.36666666666666664" />
         <entry key="..\:/Users/fatim/AndroidStudioProjects/listary/app/src/main/res/layout/activity_pentry.xml" value="0.36666666666666664" />
         <entry key="..\:/Users/fatim/AndroidStudioProjects/listary/app/src/main/res/layout/activity_register.xml" value="0.36666666666666664" />
+        <entry key="..\:/Users/fatim/AndroidStudioProjects/listary/app/src/main/res/layout/activity_register_product.xml" value="0.358695652173913" />
         <entry key="..\:/Users/fatim/AndroidStudioProjects/listary/app/src/main/res/layout/login_activity.xml" value="0.36666666666666664" />
         <entry key="..\:/Users/phtar/AndroidStudioProjects/listary-TelasAtualizadaS/app/src/main/res/layout/activity_main.xml" value="0.33" />
         <entry key="..\:/Users/phtar/Documents/GitHub/listary/app/src/main/res/layout/activity_buscar_produtos.xml" value="0.2826086956521739" />
