+++ app/src/main/res/layout/activity_register_product.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".view.RegisterProduct">
+    tools:context=".view.createProduct.RegisterProduct">
 
     <androidx.constraintlayout.widget.Guideline
         android:id="@+id/guideline8"
